requirements.txt
# Core Flask requirements
Flask==3.0.2
Flask-SocketIO==5.3.6
python-dotenv==1.0.1

# ASGI/WSGI servers
gunicorn==21.2.0
gevent==23.9.0  # Includes websocket support now

# AutoGen requirements (pinned to specific versions)
autogen-agentchat==0.4.9.2
autogen-core==0.4.9.2
autogen-ext==0.4.9.2

# Azure requirements
azure-core==1.32.0
azure-search-documents==11.5.2
azure-identity==1.21.0

# OpenAI requirements
openai==1.66.3

# Essential dependencies
greenlet==3.1.1
requests==2.32.3
urllib3==2.3.0
pydantic==2.10.6
tiktoken==0.9.0
numpy==1.24.0  # Required by some AutoGen components